+++ src/sum_of_voting.h
@@ -12,13 +12,13 @@
 using namespace std;
 
 struct PlayerWeights {
-  PlayerWeights(const vector<ll> &weights, int idx);
+  PlayerWeights(const vector<int> &weights, int idx);
   PlayerWeights() = default;
   PlayerWeights(PlayerWeights &&) = default;
   PlayerWeights(const PlayerWeights &) = default;
   PlayerWeights & operator = (const PlayerWeights &) = default;
 
-  vector<ll> weights; // TODO: optimize this with a reference
+  vector<int> weights; // TODO: optimize this with a reference
   int idx;
 
   bool operator < (const PlayerWeights & o) const;
@@ -27,7 +27,7 @@ struct PlayerWeights {
 
 class SumOfVoting: public CoalGame {
 public:
-  SumOfVoting(const vector<vector<ll>> & weights, const vector<ll> & quotas);
+  SumOfVoting(const vector<vector<int>> & weights, const vector<int> & quotas);
   /*SumOfVoting(const vector<VotingNonunique> & games): CoalGame(games[0].players), gamesNonunique(games) {
   }*/
   SumOfVoting(int players): CoalGame(players) {
@@ -41,12 +41,12 @@ public:
   vector<double> banzhafTop(int topN);
 
   // doesn't return less than numberOfTopPlayers
-  vector<int> getTopPlayers(const vector<vector<ll>> & weights, int numberOfTopPlayers);
+  vector<int> getTopPlayers(const vector<vector<int>> & weights, int numberOfTopPlayers);
 
 protected:
   vector<VotingNonunique> gamesNonunique;
   vector<VotingGame> gamesUnique;
-  vector<vector<ll>> getWeights();
+  vector<vector<int>> getWeights();
 };
